Problem: compile warnings on Windows
Solution: fix them

+++ src/zgossip_msg.c
@@ -65,7 +65,7 @@ struct _zgossip_msg_t {
 
 //  Put a 1-byte number to the frame
 #define PUT_NUMBER1(host) { \
-    *(byte *) self->needle = (host); \
+    *(byte *) self->needle = (byte) (host); \
     self->needle++; \
 }
